+++ src/api/endpoints/execute/get-execute-buy/v2.ts
@@ -10,7 +10,7 @@ import Joi from "joi";
 import { redb } from "@/common/db";
 import { logger } from "@/common/logger";
 import { slowProvider } from "@/common/provider";
-import { bn, formatEth, fromBuffer, toBuffer } from "@/common/utils";
+import { bn, formatEth, fromBuffer, regex, toBuffer } from "@/common/utils";
 import { config } from "@/config/index";
 import { Sources } from "@/models/sources";
 
@@ -21,14 +21,14 @@ export const getExecuteBuyV2Options: RouteOptions = {
   tags: ["api", "Router"],
   plugins: {
     "hapi-swagger": {
-      order: 11,
+      order: 10,
     },
   },
   validate: {
     query: Joi.object({
       token: Joi.string()
         .lowercase()
-        .pattern(/^0x[a-fA-F0-9]{40}:[0-9]+$/)
+        .pattern(regex.token)
         .description(
           "Filter to a particular token. Example: `0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63:123`"
         ),
@@ -41,14 +41,14 @@ export const getExecuteBuyV2Options: RouteOptions = {
       tokens: Joi.array().items(
         Joi.string()
           .lowercase()
-          .pattern(/^0x[a-fA-F0-9]{40}:[0-9]+$/)
+          .pattern(regex.token)
           .description(
             "Array of tokens user is buying. Example: `tokens[0]: 0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63:704 tokens[1]: 0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63:979`"
           )
       ),
       taker: Joi.string()
         .lowercase()
-        .pattern(/^0x[a-fA-F0-9]{40}$/)
+        .pattern(regex.address)
         .required()
         .description(
           "Address of wallet filling the order. Example: `0xF296178d553C8Ec21A2fBD2c5dDa8CA9ac905A00`"
@@ -56,10 +56,11 @@ export const getExecuteBuyV2Options: RouteOptions = {
       onlyQuote: Joi.boolean().default(false).description("If true, only quote will be returned."),
       source: Joi.string()
         .lowercase()
+        .pattern(regex.domain)
         .description("Filling source used for attribution. Example: `reservoir.market`"),
       referrer: Joi.string()
         .lowercase()
-        .pattern(/^0x[a-fA-F0-9]{40}$/)
+        .pattern(regex.address)
         .default(AddressZero)
         .description(
           "Wallet address of referrer. Example: `0xF296178d553C8Ec21A2fBD2c5dDa8CA9ac905A00`"
@@ -74,11 +75,14 @@ export const getExecuteBuyV2Options: RouteOptions = {
       partial: Joi.boolean()
         .default(false)
         .description("If true, partial orders will be accepted."),
+      noDirectFilling: Joi.boolean().description(
+        "If true, all fills will be executed through the router."
+      ),
       maxFeePerGas: Joi.string()
-        .pattern(/^[0-9]+$/)
+        .pattern(regex.number)
         .description("Optional. Set custom gas price."),
       maxPriorityFeePerGas: Joi.string()
-        .pattern(/^[0-9]+$/)
+        .pattern(regex.number)
         .description("Optional. Set custom gas price."),
       skipBalanceCheck: Joi.boolean()
         .default(false)
@@ -104,10 +108,8 @@ export const getExecuteBuyV2Options: RouteOptions = {
       quote: Joi.number().unsafe(),
       path: Joi.array().items(
         Joi.object({
-          contract: Joi.string()
-            .lowercase()
-            .pattern(/^0x[a-fA-F0-9]{40}$/),
-          tokenId: Joi.string().lowercase().pattern(/^\d+$/),
+          contract: Joi.string().lowercase().pattern(regex.address),
+          tokenId: Joi.string().lowercase().pattern(regex.number),
           quantity: Joi.number().unsafe(),
           source: Joi.string().allow("", null),
           quote: Joi.number().unsafe(),
@@ -360,11 +362,24 @@ export const getExecuteBuyV2Options: RouteOptions = {
         return { quote, path };
       }
 
+      // Use either the source or the old referrer
+      if (!query.source && query.referrer !== AddressZero) {
+        const source = await sources.getByAddress(query.referrer);
+        if (source) {
+          query.source = source.domain;
+        }
+      }
+
       const router = new Sdk.Router.Router(config.chainId, slowProvider);
       const tx = await router.fillListingsTx(listingDetails, query.taker, {
-        referrer: query.referrer,
-        referrerFeeBps: query.referrerFeeBps,
+        referrer: query.source,
+        fee: {
+          recipient: query.referrer,
+          bps: query.referrerFeeBps,
+        },
         partial: query.partial,
+        // Force router filling so that we don't lose any attribution
+        noDirectFilling: true,
       });
 
       // Check that the taker has enough funds to fill all requested tokens