+++ src/jobs/index.ts
@@ -23,7 +23,6 @@ import "@/jobs/token-updates";
 import "@/jobs/update-attribute";
 
 // Export all job queues for monitoring through the BullMQ UI
-// TODO: Restrict the BullMQ UI via password
 
 import * as fixActivitiesMissingCollection from "@/jobs/activities/fix-activities-missing-collection";
 import * as processActivityEvent from "@/jobs/activities/process-activity-event";
@@ -36,7 +35,9 @@ import * as backfillBlockTimestamps from "@/jobs/backfill/backfill-block-timesta
 import * as backfillFillEventsCreatedAt from "@/jobs/backfill/backfill-fill-events-created-at";
 import * as backfillFillEventsFillSource from "@/jobs/backfill/backfill-fill-events-fill-source";
 import * as backfillFillEventsOrderSource from "@/jobs/backfill/backfill-fill-events-order-source";
+import * as backfillFillEventsWashTradingScore from "@/jobs/backfill/backfill-fill-events-wash-trading-score";
 import * as backfillTransactionBlockFields from "@/jobs/backfill/backfill-transaction-block-fields";
+import * as backfillTransactions from "@/jobs/backfill/backfill-transactions";
 
 import * as topBidUpdate from "@/jobs/bid-updates/top-bid-update-queue";
 
@@ -76,6 +77,7 @@ import * as bundleOrderUpdatesByMaker from "@/jobs/order-updates/by-maker-bundle
 import * as removeBuyOrderEvents from "@/jobs/order-updates/remove-buy-order-events";
 
 import * as orderbookOrders from "@/jobs/orderbook/orders-queue";
+import * as orderbookPostOrderExternal from "@/jobs/orderbook/post-order-external";
 import * as resyncOrdersSource from "@/jobs/orderbook/resync-orders-source";
 import * as orderbookTokenSets from "@/jobs/orderbook/token-sets-queue";
 
@@ -90,8 +92,6 @@ import * as resyncAttributeFloorSell from "@/jobs/update-attribute/resync-attrib
 import * as resyncAttributeKeyCounts from "@/jobs/update-attribute/resync-attribute-key-counts";
 import * as resyncAttributeValueCounts from "@/jobs/update-attribute/resync-attribute-value-counts";
 
-import * as orderbookPostOrderExternal from "@/jobs/orderbook/post-order-external";
-
 export const allJobQueues = [
   fixActivitiesMissingCollection.queue,
   processActivityEvent.queue,
@@ -104,7 +104,9 @@ export const allJobQueues = [
   backfillFillEventsCreatedAt.queue,
   backfillFillEventsFillSource.queue,
   backfillFillEventsOrderSource.queue,
+  backfillFillEventsWashTradingScore.queue,
   backfillTransactionBlockFields.queue,
+  backfillTransactions.queue,
 
   topBidUpdate.queue,
 
@@ -144,6 +146,7 @@ export const allJobQueues = [
   removeBuyOrderEvents.queue,
 
   orderbookOrders.queue,
+  orderbookPostOrderExternal.queue,
   orderbookTokenSets.queue,
   resyncOrdersSource.queue,