+++ src/sync/events/utils.ts
@@ -4,14 +4,14 @@ import pLimit from "p-limit";
 import { baseProvider, slowProvider } from "@/common/provider";
 import { bn } from "@/common/utils";
 import { getBlocks, saveBlock } from "@/models/blocks";
+import { Sources } from "@/models/sources";
 import { getTransaction, saveTransaction } from "@/models/transactions";
-import { logger } from "@/common/logger";
 
-export const fetchBlock = async (blockNumber: number) =>
+export const fetchBlock = async (blockNumber: number, force = false) =>
   getBlocks(blockNumber)
     // Only fetch a single block (multiple ones might be available due to reorgs)
     .then(async (blocks) => {
-      if (blocks.length) {
+      if (blocks.length && !force) {
         return blocks[0];
       } else {
         const block = await baseProvider.getBlockWithTransactions(blockNumber);
@@ -63,20 +63,13 @@ export const fetchTransaction = async (txHash: string) =>
     // a good assumption so we should force re-fetch the new block
     // together with its transactions when a reorg happens.
 
-    // In order to get all transaction fields we need to make two calls:
-    // - `eth_getTransactionByHash`
-    // - `eth_getTransactionReceipt`
-
-    logger.info("debug", `Fetching tx ${txHash}`);
-
     let tx = await baseProvider.getTransaction(txHash);
     if (!tx) {
       tx = await slowProvider.getTransaction(txHash);
     }
 
-    logger.info("debug", `Got tx: ${JSON.stringify(tx)}`);
-
-    const blockTimestamp = (await fetchBlock(tx.blockNumber!)).timestamp;
+    // Also fetch all transactions within the block
+    const blockTimestamp = (await fetchBlock(tx.blockNumber!, true)).timestamp;
 
     // TODO: Fetch gas fields via `eth_getTransactionReceipt`
     // Sometimes `effectiveGasPrice` can be null
@@ -96,3 +89,27 @@ export const fetchTransaction = async (txHash: string) =>
       // gasFee: txReceipt.gasUsed.mul(gasPrice).toString(),
     });
   });
+
+export const getOrderSourceByOrderKind = async (orderKind: string): Promise<number | null> => {
+  try {
+    const sources = await Sources.getInstance();
+
+    switch (orderKind) {
+      case "x2y2":
+        return sources.getByDomain("x2y2.io").id;
+      case "foundation":
+        return sources.getByDomain("foundation.app").id;
+      case "looks-rare":
+        return sources.getByDomain("looksrare.org").id;
+      case "seaport":
+      case "wyvern-v2":
+      case "wyvern-v2.3":
+        return sources.getByDomain("opensea.io").id;
+      default:
+        // For all other order kinds we cannot default the source
+        return null;
+    }
+  } catch (error) {
+    return null;
+  }
+};